refactor: :recycle: Move some functions from dict/new to dict/util
